+++ projectoEntrega1/src/projectoEntrega1/Models/WeightedEdge.java
@@ -3,9 +3,9 @@ package projectoEntrega1.Models;
 public class WeightedEdge {
 	int o;
 	int d;
-	double w;
+	int w;
 	
-	public WeightedEdge(int o, int d, double w) {
+	public WeightedEdge(int o, int d, int w) {
 		this.o = o;
 		this.d = d;
 		this.w = w;
